Connect plotting and html summary in new class
