+++ pyFAST/__init__.py
@@ -4,6 +4,7 @@
 from .case_map import CASE_MAP
 from .executor import Executor
 from .regression_tester import RegressionTester
+from .postprocessor import SummaryHandler
 
 import os
 ROOT = os.path.abspath(os.path.dirname(__file__))
