+++ pyFAST/utilities/__init__.py
@@ -3,5 +3,4 @@
 
 from .norm import calculate_norms, pass_regression_test
 from .fast_io import load_output
-from .plotting import plot_error, create_case_summary
 from .utilities import *
